# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Laravel
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://127.0.0.1:8000'
paths:
...
  '/api/users/{id}':
    get:
      summary: 'Get user by id'
      operationId: getUserById
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    email: admin@gmail.com
                    email_verified_at: null
                    created_at: '2023-09-27T01:43:12.000000Z'
                    updated_at: '2023-09-27T01:43:12.000000Z'
                    user_name: admin
                    full_name: Admin
                    phone: '09331213344'
                    created_by: 1
                    updated_by: 1
                    role_id: 1
                    department_id: 1
                    is_super_admin: false
                    is_teacher: false
                    status: enable
                    social_id: 1
                    social_provider: google
                  message: success
                  status_code: 200
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: admin@gmail.com
                      email_verified_at:
                        type: string
                        example: null
                      created_at:
                        type: string
                        example: '2023-09-27T01:43:12.000000Z'
                      updated_at:
                        type: string
                        example: '2023-09-27T01:43:12.000000Z'
                      user_name:
                        type: string
                        example: admin
                      full_name:
                        type: string
                        example: Admin
                      phone:
                        type: string
                        example: '09331213344'
                      created_by:
                        type: integer
                        example: 1
                      updated_by:
                        type: integer
                        example: 1
                      role_id:
                        type: integer
                        example: 1
                      department_id:
                        type: integer
                        example: 1
                      is_super_admin:
                        type: boolean
                        example: false
                      is_teacher:
                        type: boolean
                        example: false
                      status:
                        type: string
                        example: enable
                      social_id:
                        type: integer
                        example: 1
                      social_provider:
                        type: string
                        example: google
                  message:
                    type: string
                    example: success
                  status_code:
                    type: integer
                    example: 200
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: unauthorized
                  status_code: 401
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: unauthorized
                  status_code:
                    type: integer
                    example: 401
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: internal_server_error
                  status_code: 500
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: internal_server_error
                  status_code:
                    type: integer
                    example: 500
      tags:
        - 'User management'
    patch:
      summary: 'Update user'
      operationId: updateUser
      description: ''
      parameters: []
      responses:
        200:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data:
                    id: 1
                    email: admin@gmail.com
                    email_verified_at: null
                    created_at: '2023-09-27T01:43:12.000000Z'
                    updated_at: '2023-09-27T01:43:12.000000Z'
                    user_name: admin
                    full_name: Admin
                    phone: '09331213344'
                    created_by: 1
                    updated_by: 1
                    role_id: 1
                    department_id: 1
                    is_super_admin: false
                    is_teacher: false
                    status: enable
                    social_id: 1
                    social_provider: google
                  message: success
                  status_code: 200
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      email:
                        type: string
                        example: admin@gmail.com
                      email_verified_at:
                        type: string
                        example: null
                      created_at:
                        type: string
                        example: '2023-09-27T01:43:12.000000Z'
                      updated_at:
                        type: string
                        example: '2023-09-27T01:43:12.000000Z'
                      user_name:
                        type: string
                        example: admin
                      full_name:
                        type: string
                        example: Admin
                      phone:
                        type: string
                        example: '09331213344'
                      created_by:
                        type: integer
                        example: 1
                      updated_by:
                        type: integer
                        example: 1
                      role_id:
                        type: integer
                        example: 1
                      department_id:
                        type: integer
                        example: 1
                      is_super_admin:
                        type: boolean
                        example: false
                      is_teacher:
                        type: boolean
                        example: false
                      status:
                        type: string
                        example: enable
                      social_id:
                        type: integer
                        example: 1
                      social_provider:
                        type: string
                        example: google
                  message:
                    type: string
                    example: success
                  status_code:
                    type: integer
                    example: 200
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: unauthorized
                  status_code: 401
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: unauthorized
                  status_code:
                    type: integer
                    example: 401
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: internal_server_error
                  status_code: 500
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: internal_server_error
                  status_code:
                    type: integer
                    example: 500
      tags:
        - 'User management'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: hmh
                full_name:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: mfg
                email:
                  type: string
                  description: 'Must be a valid email address. Must not be greater than 255 characters.'
                  example: charity.lockman@example.org
                phone:
                  type: string
                  description: 'Must not be greater than 255 characters.'
                  example: mwfqprsotqkvichs
                role_id:
                  type: integer
                  description: ''
                  example: 9
                department_id:
                  type: integer
                  description: ''
                  example: 15
                is_super_admin:
                  type: boolean
                  description: ''
                  example: true
                is_teacher:
                  type: boolean
                  description: ''
                  example: false
              required:
                - user_name
                - full_name
                - email
    delete:
      summary: 'Delete user'
      operationId: deleteUser
      description: ''
      parameters: []
      responses:
        204:
          description: ''
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: unauthorized
                  status_code: 401
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: unauthorized
                  status_code:
                    type: integer
                    example: 401
        500:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  data: null
                  message: internal_server_error
                  status_code: 500
                properties:
                  data:
                    type: string
                    example: null
                  message:
                    type: string
                    example: internal_server_error
                  status_code:
                    type: integer
                    example: 500
      tags:
        - 'User management'
    parameters:
      -
        in: path
        name: id
        description: 'The ID of the user.'
        example: 1
        required: true
        schema:
          type: integer
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: 'You can retrieve your token by visiting your dashboard and clicking <b>Generate API token</b>.'
security:
  -
    default: []